add small screen support for deals and related transactions
