+++ src/containers/Deals.js
@@ -2,19 +2,19 @@ import { useContext, useEffect, useState } from 'react';
 // import { TailSpin } from 'react-loader-spinner';
 import DashboardContext from '../contexts/DashboardContext';
 import Deal from '../components/Deal';
-import CreateDeal from '../components/CreateDeal';
+// import CreateDeal from '../components/CreateDeal';
 import DealTransactions from './DealTransactions';
-import CreateTransaction from '../components/CreateTransaction';
+// import CreateTransaction from '../components/CreateTransaction';
 import EditTransaction from '../components/EditTransaction';
 import { filterDealTransactions, fetchDealName } from '../helper_functions/helperMethods';
 import DeleteWarning from '../layouts/DeleteWarning';
 import '../styles/Deals.css';
-import ExcelExport from '../components/ExportExcel';
+// import ExcelExport from '../components/ExportExcel';
 
 const Deals = () => {
   const { deals, showEditTransactionForm, txs } = useContext(DashboardContext);
-  const [showCreateDeal, setShowCreateDeal] = useState(false);
-  const [showCreateTransaction, setShowCreateTransaction] = useState(false);
+  // const [showCreateDeal, setShowCreateDeal] = useState(false);
+  // const [showCreateTransaction, setShowCreateTransaction] = useState(false);
   const [dealTransacts, setDealTransacts] = useState([]);
   const [dealName, setDealName] = useState('');
   const [showDeleteWarning, setShowDeleteWarning] = useState(false);
@@ -25,14 +25,14 @@ const Deals = () => {
     setDealTransacts(filterDealTransactions(txs, dealID));
     setDealName(fetchDealName(dealID, deals));
   }, [txs, dealID, deals]);
-  const handleCreateDeal = (e) => {
-    setShowCreateDeal(!showCreateDeal);
-    e.preventDefault();
-  };
-  const handleCreateTransaction = (e) => {
-    setShowCreateTransaction(!showCreateTransaction);
-    e.preventDefault();
-  };
+  // const handleCreateDeal = (e) => {
+  //   setShowCreateDeal(!showCreateDeal);
+  //   e.preventDefault();
+  // };
+  // const handleCreateTransaction = (e) => {
+  //   setShowCreateTransaction(!showCreateTransaction);
+  //   e.preventDefault();
+  // };
   return (
     <div className="deals-container">
       <h3 className="deal-transactions-header">Deals related Transactions</h3>
@@ -61,17 +61,19 @@ const Deals = () => {
       <div className="transactions-container">
         <DealTransactions dealTransacts={dealTransacts} dealName={dealName} />
       </div>
-      <button type="submit" onClick={handleCreateDeal} className="create-deal">Create Deal</button>
-      <button type="submit" onClick={handleCreateTransaction} className="create-transaction">Enter Transaction</button>
+      {/* <button type="submit" onClick={handleCreateDeal}
+      className="create-deal">Create Deal</button>
+      <button type="submit" onClick={handleCreateTransaction}
+      className="create-transaction">Enter Transaction</button>
       <div className="excel-export">
         <ExcelExport excelData={txs} fileName={(new Date()).toISOString()} />
-      </div>
-      {showCreateDeal && <CreateDeal setShowCreateDeal={setShowCreateDeal} />}
+      </div> */}
+      {/* {showCreateDeal && <CreateDeal setShowCreateDeal={setShowCreateDeal} />}
       {showCreateTransaction && (
       <CreateTransaction
         setShowCreateTransaction={setShowCreateTransaction}
       />
-      )}
+      )} */}
       { showEditTransactionForm && <EditTransaction /> }
       {showDeleteWarning && <DeleteWarning fn={setShowDeleteWarning} dealID={dealID} />}
     </div>
