+++ src/styles/App.css
@@ -1,3 +1,6 @@
 /* .App {
   background-color: rgba(165, 230, 230, 0.555);
 } */
+::-webkit-scrollbar {
+  width: 0;
+}
