+++ src/styles/Balance.css
@@ -27,7 +27,7 @@
     justify-content: space-around;
     color: aliceblue;
     font-weight: bolder;
-    top: 5vh;
+    top: 5.5vh;
     left: 10vw;
     align-self: center;
   }
