+++ src/styles/CreateTransaction.css
@@ -9,7 +9,7 @@
 }
 
 @media only screen and (max-width: 428px) {
-  .create-deal-container {
+  /* .create-deal-container {
     width: 90%;
-  }
+  } */
 }
