+++ src/styles/Dashboard.css
@@ -38,6 +38,33 @@
     display: flex;
     flex-direction: row;
   }
+
+  .create-deal,
+  .create-transaction {
+    width: 20%;
+    margin: auto 25px;
+    height: 5vh;
+    color: white;
+    background-color: green;
+    font-size: larger;
+    font-weight: bolder;
+    border-radius: 10px;
+    cursor: pointer;
+  }
+
+  .create-deal {
+    position: fixed;
+    bottom: 0%;
+    left: 0%;
+    z-index: 999;
+  }
+
+  .create-transaction {
+    position: fixed;
+    bottom: 0%;
+    left: 40%;
+    z-index: 999;
+  }
 }
 
 @media only screen and (max-width: 428px) {
@@ -49,7 +76,7 @@
     display: flex;
     flex-direction: column;
     position: fixed;
-    top: -6vh;
+    top: -5.5vh;
   }
 
   .deals-transactions-switch {
@@ -64,7 +91,30 @@
     background-color: greenyellow;
   }
 
-  /* .deal-transactions-container {
+  .create-deal,
+  .create-transaction {
+    width: 30%;
+    margin: auto 10px;
+    height: 3vh;
+    color: white;
+    background-color: green;
+    font-size: 12px;
+    font-weight: bolder;
+    border-radius: 10px;
+    cursor: pointer;
+  }
 
-  } */
+  .create-deal {
+    position: fixed;
+    bottom: 0%;
+    left: 0%;
+    z-index: 999;
+  }
+
+  .create-transaction {
+    position: fixed;
+    bottom: 0%;
+    left: 35%;
+    z-index: 999;
+  }
 }
