+++ src/styles/Deal.css
@@ -1,65 +1,131 @@
-.deal-container {
-  width: 96%;
-  margin: 1px;
-  padding: 5px;
-  background-color: white;
+@media only screen and (min-width: 429px) {
+  .deal-container {
+    width: 96%;
+    margin: auto;
+    padding: 5px;
+    background-color: white;
+  }
+
+  .deal-container > button {
+    width: 100%;
+    border-radius: 10px;
+  }
+
+  .deal-container > button:hover {
+    background-color: aquamarine;
+  }
+
+  .deal-row {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    width: 100%;
+  }
+
+  .deal-btn {
+    display: flex;
+    flex-direction: column;
+  }
+
+  .deal-content {
+    display: flex;
+    flex-direction: row;
+    justify-content: left;
+    width: 100%;
+  }
+
+  .deal-content > p {
+    margin-bottom: 0;
+  }
+
+  .deal-head {
+    justify-content: space-around;
+    align-items: center;
+    font-weight: bold;
+  }
+
+  .deal-balance p {
+    align-content: center;
+    margin: 5px 0;
+  }
+
+  .deal-content p:last-child {
+    font-weight: bolder;
+    margin-left: 5px;
+  }
+
+  .deal-balance {
+    justify-content: center !important;
+    align-self: center;
+    font-size: large;
+    border: 1px solid black;
+    border-radius: 5px;
+    margin: 5px 0;
+    background-color: green;
+    color: white;
+  }
 }
 
-.deal-container > button {
-  width: 100%;
-  border-radius: 10px;
-}
+@media only screen and (max-width: 428px) {
+  .deal-container {
+    padding: 5px;
+    background-color: white;
+  }
 
-.deal-container > button:hover {
-  background-color: aquamarine;
-}
+  .deal-container > button:hover {
+    background-color: aquamarine;
+  }
 
-.deal-row {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 100%;
-}
+  .deal-row {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    width: 100%;
+  }
 
-.deal-btn {
-  display: flex;
-  flex-direction: column;
-}
+  .deal-btn {
+    display: flex;
+    flex-direction: column;
+    width: 90vw;
+    border-radius: 10px;
+    background-size: cover;
+  }
 
-.deal-content {
-  display: flex;
-  flex-direction: row;
-  justify-content: left;
-  width: 100%;
-}
+  .deal-content {
+    display: flex;
+    flex-direction: row;
+    justify-content: left;
+    width: 100%;
+  }
 
-.deal-content > p {
-  margin-bottom: 0;
-}
+  .deal-content > p {
+    margin-bottom: 0;
+  }
 
-.deal-head {
-  justify-content: space-around;
-  align-items: center;
-  font-weight: bold;
-}
+  .deal-head {
+    justify-content: space-around;
+    align-items: center;
+    font-weight: bold;
+  }
 
-.deal-balance p {
-  align-content: center;
-  margin: 5px 0;
-}
+  .deal-balance p {
+    align-content: center;
+    margin: 5px 0;
+  }
 
-.deal-content p:last-child {
-  font-weight: bolder;
-  margin-left: 5px;
-}
+  .deal-content p:last-child {
+    font-weight: bolder;
+    margin-left: 5px;
+  }
 
-.deal-balance {
-  justify-content: center !important;
-  align-self: center;
-  font-size: large;
-  border: 1px solid black;
-  border-radius: 5px;
-  margin: 5px 0;
-  background-color: green;
-  color: white;
+  .deal-balance {
+    justify-content: center !important;
+    align-self: center;
+    font-size: large;
+    border: 1px solid black;
+    border-radius: 5px;
+    margin: 5px 0;
+    background-color: green;
+    color: white;
+  }
 }
