+++ src/styles/ExcelExport.css
@@ -1,15 +1,35 @@
-.excel-export {
-  width: 15%;
-  margin: auto 25px;
-  height: 5vh;
-  color: white;
-  background-color: green;
-  font-size: large;
-  font-weight: bold;
-  border-radius: 10px;
-  cursor: pointer;
-  position: fixed;
-  bottom: 0%;
-  left: 80%;
-  z-index: 999;
+@media only screen and (min-width: 429px) {
+  .excel-export {
+    width: 15%;
+    margin: auto 25px;
+    height: 5vh;
+    color: white;
+    background-color: green;
+    font-size: large;
+    font-weight: bold;
+    border-radius: 10px;
+    cursor: pointer;
+    position: fixed;
+    bottom: 0%;
+    left: 80%;
+    z-index: 999;
+  }
+}
+
+@media only screen and (max-width: 428px) {
+  .excel-export {
+    width: 25%;
+    margin: auto 10px;
+    height: 3vh;
+    color: white;
+    background-color: green;
+    font-size: 12px;
+    font-weight: bold;
+    border-radius: 10px;
+    cursor: pointer;
+    position: fixed;
+    bottom: 0%;
+    left: 70%;
+    z-index: 999;
+  }
 }
