+++ src/styles/SideTransactions.css
@@ -63,6 +63,10 @@
     overflow-y: scroll;
     width: 100%;
   }
+
+  /*
+  .side-transactions-table-body::-webkit-scrollbar-corner {
+  }   */
 }
 
 @media only screen and (max-width: 428px) {
@@ -90,6 +94,7 @@
   .side-transactions-table-head-content {
     display: flex;
     flex-direction: row;
+    justify-content: space-between;
     border-bottom: 1px solid blueviolet;
     border-top: 1px solid blueviolet;
     width: 100%;
@@ -117,7 +122,7 @@
   }
 
   .edit-delete {
-    width: 10vw;
+    width: 14vw;
   }
 
   .side-transactions-table-head {
@@ -130,9 +135,14 @@
   .side-transactions-table-body {
     top: 15vh;
     overflow-y: scroll;
+    height: 75vh;
     width: 100%;
 
     /* margin-bottom: 10px; */
+  }
+
+  .side-transactions-table-body::-webkit-scrollbar {
+    display: none;
   }
 
   .side-transactions-table-body p {
