+++ src/styles/Transaction.css
@@ -52,6 +52,7 @@
     display: flex;
     flex-direction: row;
     margin-left: 2vw;
+    justify-content: space-between;
   }
 
   .side-transactions-row2 {
